% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bioset.R
\docType{package}
\name{bioset-package}
\alias{bioset}
\alias{bioset-package}
\title{Convert a matrix of raw values into nice and tidy data.}
\description{
\link{bioset} is intended to save you from commons tasks when dealing with
raw data obtained e.g. from a measuring device.
}
\details{
When importing values with \link{bioset} start with \code{\link[=set_read]{set_read()}}. To
get an example for a workflow refer to the introductory vignette
(\code{vignette("introduction", "bioset")}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/randomchars42/bioset}
  \item Report bugs at \url{https://github.com/randomchars42/bioset/issues}
}

}
\author{
\strong{Maintainer}: Eike Christian Kühn \email{eike.kuehn@pixelwoelkchen.de}

}
