% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{calc_concentrations}
\alias{calc_concentrations}
\title{Bla.}
\usage{
calc_concentrations(data, cal_names, cal_values, cal_unit, target_unit,
  col_names = name, col_values = values, col_target = conc,
  col_real = real, model_func = fit_lnln, interp_func = interp_lnln)
}
\arguments{
\item{data}{A tibble containing the data.}

\item{cal_names}{A vector of strings containing the names of the samples used
as calibrators,}

\item{cal_values}{A numeric vector with the known concentrations of those
samples (must be in the same order).}

\item{cal_unit}{A string indicating the unit, see \code{\link{calc_factor}}.}

\item{target_unit}{A string indicating the unit, see
\code{\link{calc_factor}}.}

\item{col_names}{A string indicating the name of the column used to identify
the calibrators.}

\item{col_values}{A string indicating the name of the column holding the raw
values.}

\item{col_target}{A string indicating the name of the column that will be
created for the calculated concentration.}

\item{col_real}{A string indicating the name of the column that will be
created for the calculated concentration.}
}
\value{
A tibble containing all original and additional columns.
}
\description{
Vla.
}
