% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.R
\name{convert_prefix}
\alias{convert_prefix}
\title{Convert a value of the given concentration into another concentration.}
\usage{
convert_prefix(x, from, to)
}
\arguments{
\item{x}{The value to convert.}

\item{from}{A string containing the prefixed unit A.}

\item{to}{A string containing the prefixed unit B.}
}
\value{
The converted value.
}
\description{
A convenience wrapper around \code{\link{calc_factor_prefix}}.
}
\details{
Convert, e.g. "kg" to "µg". You can convert ".g", ".l", ".mol", ".M", "m^3"
(cubic metres), where "." symbolises a metric prefix:

For g, l, mol and M: m (milli), µ (micro), n (nano), p (pico) and f (femto).

For g you might use k (kilo) as well.

For m^3 (cubic metres): d (deci), c (centi) and m (milli)

"." symbolises a metric prefix (see \code{\link{calc_factor_prefix}}) :

For g, l, mol and M: m (milli), µ (micro), n (nano), p (pico) and f (femto).

For g you might use k (kilo) as well.

For m^3 (cubic metres): d (deci), c (centi) and m (milli)
}
\seealso{
Other calc: \code{\link{calc_factor_conc}},
  \code{\link{calc_factor_prefix}},
  \code{\link{convert_conc}}
}
