% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_linear}
\alias{fit_linear}
\title{Fit a linear function.}
\usage{
fit_linear(x, y)
}
\arguments{
\item{x}{The x coordinates of the points.}

\item{y}{The y coordinates of the points.}
}
\value{
Line model.
}
\description{
Fit a linear function to the points with the given x and y coordinates.
}
