% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_lnln}
\alias{fit_lnln}
\title{Fit a linear function to ln-ln transformed values.}
\usage{
fit_lnln(x, y)
}
\arguments{
\item{x}{The x coordinates of the points.}

\item{y}{The y coordinates of the points.}
}
\value{
Line model.
}
\description{
Fit a linear function to the points with the given x and y coordinates after
transforming both coordinates using ln.

Can be used in \code{\link[=set_calc_concentrations]{set_calc_concentrations()}} to fit a model to the
calibrators.
}
\seealso{
\code{\link[=set_calc_concentrations]{set_calc_concentrations()}}

Other ln-ln transformation model functions: \code{\link{interpolate_lnln}},
  \code{\link{plot_lnln}}
}
