% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{interpolate_linear}
\alias{interpolate_linear}
\title{Interpolate using a linear function.}
\usage{
interpolate_linear(y, model)
}
\arguments{
\item{y}{The y coordinates of the points.}

\item{model}{The line model.}
}
\value{
Vector of x values.
}
\description{
Interpolate values using a linear function and the coefficients
of the given line model.

Can be used in \code{\link[=set_calc_concentrations]{set_calc_concentrations()}} to interpolate
concentrations.
}
\seealso{
\code{\link[=set_calc_concentrations]{set_calc_concentrations()}}

Other linear model functions: \code{\link{fit_linear}},
  \code{\link{plot_linear}}
}
