% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{plot_linear}
\alias{plot_linear}
\title{Fit a linear function and plot it.}
\usage{
plot_linear(x, y)
}
\arguments{
\item{x}{The x coordinates of the points.}

\item{y}{The y coordinates of the points.}
}
\value{
Plot.
}
\description{
Fit a linear function to the points with the given x and y coordinates and
plot it.

Can be used after \code{\link[=set_calc_concentrations]{set_calc_concentrations()}} to visualise the fit
of the function.
}
\seealso{
\code{\link[=set_calc_concentrations]{set_calc_concentrations()}}

Other linear model functions: \code{\link{fit_linear}},
  \code{\link{interpolate_linear}}
}
