% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{plot_lnln}
\alias{plot_lnln}
\title{Fit a linear function to ln-ln transformed values and plot it.}
\usage{
plot_lnln(x, y)
}
\arguments{
\item{x}{The x coordinates of the points.}

\item{y}{The y coordinates of the points.}
}
\value{
Plot.
}
\description{
Fit a linear function to the points with the given x and y coordinates after
transforming both coordinates using ln and plot them.

Can be used after \code{\link[=set_calc_concentrations]{set_calc_concentrations()}} to visualise the fit
of the function.
}
\seealso{
\code{\link[=set_calc_concentrations]{set_calc_concentrations()}}

Other ln-ln transformation model functions: \code{\link{fit_lnln}},
  \code{\link{interpolate_lnln}}
}
