% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_calc_concentrations}
\alias{set_calc_concentrations}
\title{Calculate concentrations for the set using contained calibrators.}
\usage{
set_calc_concentrations(data, cal_names, cal_values,
  col_names = eval(parse(text = "name")), col_values = values,
  col_target = conc, col_real = real, col_recov = recovery,
  model_func = fit_lnln, interpolate_func = interpolate_lnln)
}
\arguments{
\item{data}{A tibble containing the data.}

\item{cal_names}{A vector of strings containing the names of the samples used
as calibrators,}

\item{cal_values}{A numeric vector with the known concentrations of those
samples (must be in the same order).}

\item{col_names}{The name of the column where the \code{cal_names} can be
found.}

\item{col_values}{The name of the column holding the raw values.}

\item{col_target}{The name of the column to created for the calculated
concentration.}

\item{col_real}{The name of the column to create for the known
concentrations.}

\item{col_recov}{The name of the column to create for the recovery of the
calibrators.}

\item{model_func}{A function generating a model to fit the calibrators,
e.g. \code{\link{fit_linear}}, \code{\link{fit_lnln}}}

\item{interpolate_func}{A function used to interpolate the concentrations of
the other samples, based on the model, e.g.
\code{\link{interpolate_linear}}, \code{\link{interpolate_lnln}}}
}
\value{
A tibble containing all original and additional columns.
}
\description{
If the data set is generated, for example by reading extinction rates or
relative light units from a plate, these raw values can be converted to
concentrations using data fields with known concentrations (calibrators).
}
\details{
If the data set contains samples with known concentrations (calibrators)
those can be used to interpolate the concentrations of the other samples.
}
