% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{set_calc_variability}
\alias{set_calc_variability}
\title{Calculate paramters of variability for a given set of values.}
\usage{
set_calc_variability(data, ids, ...)
}
\arguments{
\item{data}{A tibble containing the data.}

\item{ids}{The column holding the names used to group the values.}

\item{...}{The name(s) of the columns used to calculate the variability.}
}
\value{
A tibble containing all original and additional columns
  (NAMEA_mean, NAMEA_n, NAMEA_sd, NAMEA_cv, (NAMEB_mean)).
}
\description{
Calculate mean, standard deviation and coefficient of variation for groups of
values.
}
\details{
Dealing with measured values, the measurement of sample "A" is often done in
duplicates / triplicates / ... . This function groups all samples with the
same name and calculates mean, standard deviation and coefficient of
variation (= sd / mean).
}
