% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{fit_linear}
\alias{fit_linear}
\title{Fit a linear function.}
\usage{
fit_linear(x, y)
}
\arguments{
\item{x}{The x coordinates of the points.}

\item{y}{The y coordinates of the points.}
}
\value{
Line model.
}
\description{
Fit a linear function to the points with the given x and y coordinates.

Can be used in \code{\link[=set_calc_concentrations]{set_calc_concentrations()}} to fit a model to the
calibrators.
}
\seealso{
\code{\link[=set_calc_concentrations]{set_calc_concentrations()}}

Other linear model functions: \code{\link{interpolate_linear}},
  \code{\link{plot_linear}}
}
