% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{interpolate_lnln}
\alias{interpolate_lnln}
\title{Interpolate using a linear function on ln-ln transformed values.}
\usage{
interpolate_lnln(y, model)
}
\arguments{
\item{y}{The y coordinates of the points.}

\item{model}{The line model.}
}
\value{
Vector of x values.
}
\description{
Interpolate values using a linear function and the coefficients
of the given line model after transforming both coordinates using ln.

Can be used in \code{\link[=set_calc_concentrations]{set_calc_concentrations()}} to interpolate
concentrations.
}
\seealso{
\code{\link[=set_calc_concentrations]{set_calc_concentrations()}}

Other ln-ln transformation model functions: \code{\link{fit_lnln}},
  \code{\link{plot_lnln}}
}
